Commits inputs to database after starting sim

+++ src/main/java/application/view/ISimulatorGUI.java
@@ -3,7 +3,7 @@ package application.view;
 public interface ISimulatorGUI {
 	
 	// Kontrolleri tarvitsee syötteitä, jotka se välittää Moottorille
-	public double getSimulationTimeValue();
+	public int getSimulationTimeValue();
 
 	double getMeanTrashAmtPerThrow();
 	int getGarbageTruckArrivalInterval();
